don't target unassigned storages for reconciliation
To support variable replication factor, repository host node
assignments were introduced in 243dc385ccfd9f759ddb50e3fb659fc072452203.
Repository's assignments indicate which storages are supposed to
contain a copy of the repository. Since only the assigned storages
need to contain the repository, the reconciler should take the assignments
in to account.

This commit changes the reconciler to not target outdated, unassigned
repositories for reconciliation. They may still act as the source node in
order to facilitate moving repositories to different storages.

While the unassigned repositories are not targeted for now, in the future
the reconciler will delete them. This is left for later when the
functionality is more mature.

The functionality is behind a toggle as Praefect doesn't yet fully support
variable replication factor. The functionality can later be enabled with
per repository elector once the rest of the functionality has been
implemented.