fetch host node assignments from the database
In preparation for variable replication factor, Praefect gained support
for assigning host nodes for repositories in 243dc385.

The PerRepositoryRouter uses host assignments to determine which nodes
should participate in transactions or get replicated to a given write.

This commit adds a Postgres implementation of AssignmentGetter interface.
This allows the PerRepositoryRouter to get host node assignments from
the database to only include assigned nodes in transactions and
replications.