testhelper: Have Configure() return cleanup function
During tests we're accumulating quite some global data, some of which is
hard to clean up separately. As we don't have any shared global test
directory in which we can put all generated test data, they're thus
all over `/tmp` and never get cleaned up. As a result, after a few runs
the temporary directory will be cluttered by thousands of files and
directories generated by our test suite.

This commit is a preparation for such a global unified test directory,
which is going to be set up by `testhelper.Configure()`. In order to
assure it'll get cleaned up, we add a (still empty) cleanup function and
adjust all callers to invoke it. This function will at a later point be
used to delete the temporary test directory.