tests: Unify test setup
There's been some proliferation in our test's main functions. Naturally
enough, this causes some of them to not work correctly, e.g. because
they use `defer` in the `TestMain` function, which is not getting
executed if `os.Exit` is called.

Unify existing setup functions to follow the same pattern, which is that
`TestMain()` calls out to `testMain()`.