add a linter for checking sql row iterating error is checked
Adds a linter for checking the iteration error of SQL rows is checked.

Small section of code had to be modified to pass the linter but this
is a minor change compared to missing any error checks. Additionally
the affected code would fail if there was a problem closing the rows
which is unnecessary as we already got the data we needed.