featureflag: Add LogCommandStats to the set of all flags
The `LogCommandStats` feature flag is currently not part of the set of
all feature flags, which means that we don't actually exercise it if
feature sets are being used. Adding it in used to cause segfaults in our
test suite, but that got fixed with the preceding commit. So let's add
it to that set to increase its test exposure.