smarthttp: Fix concurrent access to logrus metadata
Go's map types have some safety measures to detect concurrent reads and
writes to the same map, which will cause a panic. The ctxlogrus logger
is backed by a map, which thus means that we aren't allowed to extract
it from context while concurrently adding any fields to it. But we're
triggering such a race smarthttp.PostUploadPack, even though this is
quite hidden:

    1. We're spawning a Goroutine which is responsible for parsing the
       packfile negotiation. This Goroutine will log an error in case
       parsing failed. But as we're not waiting for this Goroutine to
       exit in case the spawned git-upload-pack(1) executable exited
       successfully, it may leak the parent's function context. This is
       the read side.

    2. The commandstatshandler will add process metrics to the context
       logger after any RPC has finished. This is the write side.

As the Goroutine is leaking, it may execute code at the same time as the
commandstatshandler does, which then leads to the panic.

Fix the issue by always waiting for the Goroutine to exit.