Port UserDeleteTag to Go
The code is initially derived from [1], but mainly draws on my own
subsequent fixes to the User{Create,Delete}Branch code. See [2][3]
etc. for those fixes.

As with the Branch code the aim is to retain bug-for-bug compatibility
with the existing Ruby responses. Some of them don't make much sense,
but we can change that later once we're migrated to the Go code.

1. c1e3ccca ("Initial go implementation of UserCreateBranch",
   2020-09-30)

2. 816ad5b23 (User{Branch,Submodule}: remove erroneously copy/pasted
   error handling, 2020-11-24)

3. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2864