helper: Convert to use `testhelper.NewTestRepo()`
We're about to move the test storage into the global temporary test
directory, which will make use of `testhelper.TestRepository()` not work
anymore. This commit thus refactors `internal/helper` to use
`testhelper.NewTestRepo()` instead.

There is one non-trivial conversion related to relative paths and
escaping storages. As previously the repository was directly located in
the storage root, it was trivial to escape that storage with only
"/../..". But now the test repository is located like it typically would
be in hashed storage with sharding directories, which requires us to
traverse further in order to trigger the same error.