extract a test suite for repository
This commit extracts a test suite for testing implementations of
`ìnternal/git.Repository`. Extracting allows us to share the tests
between different implementations to avoid duplicating the code and
to make sure the implementations behave the same.

The test suite is exported in order to have it be callable from
other packages. This allows us to move the implementations in to
subpackages in order to not pull their dependencies in to the high
level git package. In our case, this allows us to avoid dependency
cycles when testing services which depend on the git package with
implementations that depend on some services.