move remote repository to a subpackage
RemoteRepository pulls in dependencies which are unrelated to git
to the git package. This makes testing also more involved as we
remote repository depends on multiple Gitaly services for its tests,
which some depend on git package. This causes a dependency cycle
forcing git packages tests to stay in a separate test package.

This commit moves remote repository to its own subpackage to remove
the dependency cycle and fixes the few call sites to use the new
package. This way git package can stay agnostic to whatever
implementations of Repository interface might exist.