move git package's tests from git_test back to git
In 99e95ce5, the RemoteRepository was added to the git package.
RemoteRepository's tests created a cyclic dependency as they depend
on Gitaly services which depend back on the git package. The tests
were moved to git_test package to avoid the problem.

Since then, the RemoteRepository has been moved to a subpackage which
has removed the cyclic dependency between the various packages. This
commit moves git packages test back to the git package.