repository: Fix archive LFS test writing into repo
The testcase `TestGetArchiveWithLfsSuccess()` is setting up a GitLab
server together with gitlab-shell. But as `SetupAndStartGitlabServer`
depends on the global Gitaly config in order to determine where the
gitlab-shell directory is, and that config value isn't yet set up when
the function is called, the secret ends up in a path relative to the
current working directory.

Fix the issue by modifying the global configuration before calling
`SetupAndStartGitlabServer()`.