updateReferenceWithHooks: fix hook std(err|out) handling to be like Ruby
In the Ruby code in GitLab::Git::Hook we return either stderr if
there's anything there, or fall back to stdout.

This behavior dates back to gitlab-org/gitlab@926a8ab4765 ("Handle
custom Git hook result in GitLab UI", 2016-07-04). The current Ruby
code Gitaly runs was copy/pasted into the tree in 5d0dc144 ("Implement
CommitService.FindCommits", 2017-08-14).

When this was ported to Go in f7bb6743 ("operations: Port
UserMergeBranch from Ruby to Go", 2020-09-09) this logic error of
always using stdout was introduced. Fix it.

This makes the test added in the previous commit work. I'm fixing this
to be bug-compatible with the Ruby implementation which uses/used
".blank?".

I asked @vsizov whether he could recall whether the ".blank?" test was
needed or just some mistaken idiom for "empty string". He thought it
was safe to change it in Go to just use the len() test, but wasn't
sure.

I'm keeping the ".blank?" emulation here not so much out of the
paranoia of being bug-compatible, but because it's easier to run the
tests without things being conditionally different under the in-flight
features to migrate functionality to Go.