coordinator: Move transactional RPC map
We're about to add all of our storage-scoped mutating RPCs to this map.
In order to not miss any RPCs, there'll be a safety check which verifies
in the `init()` function that all those RPCs are covered by that map. As
a preparatory step, this commit thus moves the map further up in the
code such that it's going to be located next to the new `init()`
function.