commit: Fix ListFiles returning internal error with unborn branch
When passing an empty revision to ListFiles, then it's expected to fall
back to the default branch. This works alright as long as the repository
does have any branches, but in the edge case of a completely epmty
repository this fails. Instead, we pass the empty revision to
`repo.ContainsRef`, which returns an error. This later on gets
translated to an internal error.

Given that this is valid use of the RPC, we should instead fail with a
proper error code and error message. This commit thus converts those
failure modes to return a `FailedPrecondition` failed with an error
message stating that there's not a default branch. It also adds a bunch
of tests around unborn branches.