coordinator: Add conditions for transactional RPCs
Right now, we only either enable or disable use of transactions for RPCs
by simple presence in a map. We're about to grow more flexible here
though to discern RPCs which always use transactions, never use
transactions or conditionally use transactions based on feature flags.

This commit thus converts our map of transaction RPCs to also carry a
condition. This condition is a simple function which returns a boolean
to indicate whether a given feature shall be enabled or not.