log: Don't log hook dafa if GITALY_LOG_DIR is unset
If calling `log.NewHookLogger()`, then we'll use the environment
variable GITALY_LOG_DIR to figure out where to write the logfile to. But
in case it's unset, we end up simply writing into the current directory,
wherever that is.

Fix the issue by creating a nil-logger in case the variable is unset or
empty.