tags tests: don't inconsistently pipe GL_ID=* to stderr
Remove piping to stderr added in a7ac5fde ("Enable feature flag for go
update hooks through operations service", 2020-04-28). This made it
inconsistent with the corresponding branch test.

This removes test coverage for the stdout v.s. stderr case, that'll be
rectified in a follow-up commit which fixes a bug in that logic and
tests for it explicitly.