namespace: Stop writing data into "testdata"
We're creating a set of custom storages in our "namespace" tests. These
storages are currently created in "testdata/", but should instead be
created in a temporary directory

Fix tests to use `ioutil.TempDir()`. As we don't have a test context
available at time of creation, we cannot use `testhelper.TempDir()`.