git: Add "init" to commands which don't update refs
As we don't yet pass command options to git.SafeCmds without a repo,
using git-init(1) without `git.WithRefTxHook()` has been safe even
though it wasn't on the list of commands which don't perform any
reference updates. We're about to extend the repo-less commands though
to also take options, which is why we need to add it now.

Strictly speaking, git-init(1) does create a reference, but this really
only is the HEAD symbolic reference which points at a yet-nonexistant
branch. As the reference-transaction hook wouldn't tell us about it
anyway, it's thus still safe to treat it as non-existing. Furthermore,
our hook infrastructure expects to get a repository upon hook
invocation. But as we're just in the process of creating it, it wouldn't
be possible to provide one.

So let's just add it to the list of commands which don't do reference
updates.