test more invalid repository cases for ReplicateRepository
ReplicateRepository's tests currently do not test cases where the
source repository is invalid or where both of the repositories are
invalid. This commit extends the tests to cover these cases.

The tests expose a bug in the implementation where the returned errors
are different depending on implementation of the shelled out 'tar'.
The tests work with BSD tar but GNU tar will fail the test. This is
addressed in a follow up commit that fixes the disparity in behavior.