git: Inline `unsafeBareCmdWithoutRepo()`
The function `git.unsafeBareCmdWithoutRepo()` is a direct wrapper of
`unsafeBareCmd()`, just without an additional environment parameter.
Given that we'll have to switch its only non-test user to the variant
which does have an environment argument, let's just inline this
function. This is required in order to implement global usage of
reference-transactions.