testhelper: Add option to setup servers with internal socket
While there already is a set of tests which need to set up Gitaly's
internal socket, this set is about to grow with the move to globally
executed reference-transaction hooks. This commit thus provides a helper
option for the testserver setup which automatically creates a listener
on the internal Unix socket to make the setup easier.