gitaly-ssh: Set up hook service
In order to make use of hooks, the GRPC Hook service must be available
on Gitaly's internal socket path. There is currently only a single test
in `cmd/gitaly-ssh` which does invoke hooks, so let's just locally fix
it for that test by setting up two servers: an external one and one
which listens on the internal socket path.

This is in preparation for the reference-transaction hook being executed
globally for all reference-updating Git commands.