return a special status to indicate invalid replication source
Since the introduction of automatic reconciler, invalid repositories
which failed replication will re-attempt replication until it succeeds.
Some repositories in the database do not correspond to valid git
reppsitories on disk. Replication will never succeed for these repositories.
To avoid attempting to replicate these forever, this commit returns a special
status on these cases that allows Praefect to handle the case in a special
manner.