git: Avoid usage of hooks in tests
Our tests in `internal/git` are independent of any GRPC server setups as
they're a lot more low-level than that. As our hooks depend on the Hook
GRPC service to be accessible, the conclusion is that we cannot test
hooks in this package.

Override the hook path so that we don't try to use them. While this is
currently not an issue, this is in preparation of an extended
`WithRefTxHook()` option which also correctly sets `core.hooksPath`.