praefect: Configure hooks binary
Praefect uses a complete Gitaly server setup, which will make use of Git
hooks as soon as we start to globally enable the reference-transaction
hook for most of our Git commands. Currently, this would fail Praefect
tests though as we're not configuring the gitaly-hooks binary as a part
of our test setup.

Fix the issue by calling `testhelper.ConfigureGitalyHooksBinary()`.