simplify error tracker health checking code
Moves error tracker health checking logic in to a health check
client wrapper in order to make it transparent to the caller. This
allows us to remove the special handling of it and reuse the common
path in CheckHealth. While this simplifies the code, it also fixes
issues where health checking errors related to error thresholds
were not logged and counted like the other errors were