Tests that rely on Postgres instance are flaky
It is hard to reason about the cause of the failing PG tests, but the
failure commonly happens on attempt to truncate tables.
Logic of truncation refactored from multiple statements (per table)
to a single truncate statement with multiple tables. Potentially cause
could be a reshuffle of the statements that could cause a lock.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3208