print in-sync storages in dataloss output
`praefect dataloss` currently only prints out the outdated storages.
This makes it difficult to identify which nodes contains the up to
date copies of a repository without checking the configuration file
for the configured storages. This commits prints the in-sync replicas
in addition to the outdated ones.

This change additionally lays out the foundation to support variable
replication factor. With variable replication factor, only the storages
assigned to host a given repository should contain it. With that in mind,
it's not possible to deduce the in-sync storages by checking the
configuration.