add a mock repository store
Currently the MemoryRepositoryStore is used generally in tests. Many
of the tests do not really test repository state tracking behavior but
do need some sort of a repository store so the tested functionality
works without nil pointer panics.

This commit additionally adds a MockRepositoryStore which can be used
to mock RepositoryStore accesses when the testing something particular
about RepositoryStore's behavior. The mock defaults to behavior which
could be considered successful to avoid the need to set methods which
the user is not interested in.

The mockpaves the way for removing the MemoryRepositoryStore.
It can't be used in production as the state is not persisted.
Having to maintain an in-memory implementation of a store only for testing
add additional burder when implementing new features. Additionally, when
testing the actual store's behavior matters, it would be best served by
using the Postgres implementation as it is what eventually will run in
production.