remove MemoryRepositoryStore
This commit removes MemoryRepositoryStore. It is only used for testing
and requires additional maintenance when adding new features. Tests
would be more thorough if they were using the Postgres implementation
which is running in Production. If thoroughness is not needed but
only something that works, the DisabledRepositoryStore is sufficient
to have code pass RepositoryStore's consistency checks.

DisabledRepositoryStore replaces MemoryRepositoryStore when the
database is disabled. This should be fine as the MemoryRepositoryStore
forgets state anyway and can't be used in production.