hook: Validate input of update hook
The update hook always gets invoked with exactly three parameters which
are the reference which is to be updated as well as the old and new
object IDs. Let's verify this actually happens by checking that the
reference is non-empty and that the OIDs are valid. We could be a tad
stricter here by checking that the reference always starts with "ref/"
as it shouldn't ever be invoked for any other references, but this
commit instead errs on the safe side and only verifies that we got
something.