use postgres repository store in TestStreamDirectorMutator_Transaction
This replaces the in-memory store in Praefect's coordinator's
TestStreamDirectorMutator_Transaction tests with Postgres implementation
of the store. The test case exercises the whole mutator flow and tests
assumptions related to it. Due to this, mocking gets more involved.
As the test case covers a lot of ground, it makes sense to use the
real implementation here and perform integration testing with it.