operations: Expose server type to allow testing it
As we're about to change some implementation details of
`server.updateReferenceWithHooks()`, we want to add a set of tests which
verifies that our changes actually make sense. But the type is currently
private and thus not returned by `NewServer()`, which makes it hard to
wire up a test server.

Fix the issue by making exposing the type to other modules and returning
it from `NewServer()` directly, which allows us to test non-public
methods of that type.