gitaly-debug: has no dependency on the config file
command.GitPath() depends on the configuration provided from outside
and that configuration is not provided to the gitaly-debug executable.
The only possible case if GITALY_TESTING_GIT_BINARY env var is set to
a particular git executable. But the same could be configured using an
updated PATH.
The 'ReadIndex' function only used from the 'gitaly-debug', so it should
rely on the same 'git' binary.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699