gitaly-hooks check: check is not a gitaly hook
'check' subcommand is not a real hook and should not write to the
same log used by actual gitaly-hook subcommands.

The 'GitlabNetClient.Check' method of the gitlab-shell lib expects
to get a JSON with 'message' field in case of an error. The mock
server aligned with that expectation. As a result the error message
has another content.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699