use assignments table in per repository elector
This commit updates per repository elector to use the repository
assignments table instead of the assigned column. If any assignments
are set for a repository, the elector will only elect an assigned node
as a primary. This allows the elector to gracefully start supporting
variable replication factor on a per repository basis. Once the feature
is mature, the behavior can be further constrained.