coordinator: Categorize RPCs by whether they need transaction or not
The list of repository-scoped mutating RPCs which is consulted for
whether a given RPC shall use transactions or not currently only has two
blocks, one for unconditionally enabled respectively disabled ones. In
the next commit, we're going to conditionally enable a subset of the
disabled RPCs via feature flags.

But in fact it doesn't make any sense to set up transactions for some of
the RPCs. If we know that a given RPC won't ever update any reference,
then it's fine to not use them. Most importantly, this includes RPCs
which do repo optimizations, change config, maintain object pools and
the transaction voting RPCs themselves.

So let's group together all calls which fall into the same category as a
preparatory step to keep the actual code changes small. This shouldn't
yet trigger any change in behaviour.