git: Extract function to serialize transaction envvars
In order to make use of transactions in hooks, we require two
environment variables: one to identify transaction info and another to
identify the Praefect server. Whether to assemble them or not depends on
the presence of transaction information in the context, and only when it
does exist do we also make Praefect info available. As a result, it's a
hard error when transaction info exists, but Praefect info doesn't.

This logic is currently handled ad-hoc in `receivePackHookEnv()`. As
we'll require it also for the reference-transaction hook environment and
considering that the logic is non-trivial, this commit pulls out the
relevant code into its own new function as a preparatory commit.