implement functionality to set a repository's replication factor
This commit adds SetReplicationFactor method on the AssignmentStore
for assigning and unassigning storage nodes from a repository until
the desired replication factor is met.

Repository's assignments define where the repository should be
replicated to. Praefect already queries the assignments in the
relevant locations. If no assignments are set, Praefect replicates
the repository to every storage node. This can be wasteful as the
repository's replication factor increases with the number of storages
in the virtual storage. SetReplicationFactor allows for setting a
replication factor for a repository to support replicating to only
a specified number of storage nodes instead.