git: Pass transaction environment to reftx hooks
The reference-transaction hooks nowadays include all information
required to start transactions, except for the transaction information
itself.

This commit adds this final missing piece, which in turn will cause all
transaction-enabled RPCs to start voting on transactions for each
reference-modifying Git command. As there are only a handful of RPCs
which are currently enabled for transactions, the actual effect of this
commit is still comparatively low.